requirements.txt
Django
Markdown==2.3.1
South==0.8.2
argparse==1.2.1
distribute==0.6.24
django-custom-user==0.1
requests==0.13.2
stripe==1.9.1
wsgiref==0.1.2
